WIP Skeleton Line Lib 3.0
Wrote the skeleton library for line lib 3.0. In it you can see alot less processing will be achieved for the same amount of functionality of 2.0 with the extendablity of normal java objects finally. In 3.0 you will be working with the interfaces directly or objects depending on useage.

For now I am going to start working on SemanticVersioning and the Section API parser for the next week. When that's done I will finish this not like it's that far away from what I wrote already :)